Refactor to Shared Kernel Architecture and Pure Pydantic Models
## Refactor to Shared Kernel Architecture and Pure Pydantic Models (#79)

This commit represents a major architectural pivot, transitioning the codebase to a **Shared Kernel** pattern focused on pure Pydantic definitions and purging legacy execution-heavy logic ("zombie code").

### Core Changes

* **Architectural Refactor:** * Reorganized the project into a **Shared Kernel** structure, prioritizing clean data definitions over execution logic.
* Decoupled core models from legacy dependencies to ensure the library serves as a lightweight foundation for external services.


* **Model Reorganization:**
* Moved and refactored `models.py` into a domain-specific hierarchy under `src/coreason_manifest/definitions/`.
* Extracted topology models from recipes into standalone definitions.
* Introduced stubs for future-facing domains: `SimulationScenario`, `SimulationTrace`, `SimulationTurn`, and `AuditLog`.


* **Codebase Cleanup:**
* **Deleted obsolete source files:** Removed `server.py`, `engine.py`, `validator.py`, and their associated legacy tests.
* **Public API:** Updated `__init__.py` to provide a clean, single point of entry for the exported Shared Kernel models.


* **Tooling & Quality:**
* Added `types-PyYAML` dev dependency for improved type checking.
* Restored build pipeline and ensured **100% test coverage** for the new structure.



### Detailed File Changes

* `src/coreason_manifest/models.py` → `src/coreason_manifest/definitions/agent.py`
* `src/coreason_manifest/recipes.py` → `src/coreason_manifest/definitions/topology.py`
* New: `src/coreason_manifest/definitions/simulation.py`
* New: `src/coreason_manifest/definitions/audit.py`

https://github.com/CoReason-AI/coreason-manifest/commit/f280a8efe78f40626dc6bde54e74ba7a79c0f701
