+++ pyproject.toml
@@ -11,14 +11,8 @@ packages = [{include = "coreason_manifest", from = "src"}]
 python = ">=3.12, <3.15"
 loguru = "^0.7.2"
 pydantic = "^2.12.5"
-jsonschema = "^4.26.0"
 pyyaml = "^6.0.3"
-anyio = "^4.12.1"
 coreason-identity = "^0.4.1"
-httpx = "^0.28.1"
-aiofiles = "^23.2.1"
-fastapi = "^0.128.0"
-uvicorn = "^0.40.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^9.0.2"
@@ -29,8 +23,8 @@ mkdocs = "^1.6.1"
 mkdocs-material = "^9.7.1"
 pydantic = "^2.12.5"
 mypy = "^1.19.1"
-types-aiofiles = "^23.2.0"
 pytest-asyncio = "^1.3.0"
+types-pyyaml = "^6.0.12.20250915"
 
 [build-system]
 requires = ["poetry-core"]
