+++ requirements.txt
@@ -1,7 +1,5 @@
-aiofiles==23.2.1 ; python_version >= "3.12" and python_version < "3.15"
 annotated-doc==0.0.4 ; python_version >= "3.12" and python_version < "3.15"
 annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "3.15"
-anyio==4.12.1 ; python_version >= "3.12" and python_version < "3.15"
 attrs==25.4.0 ; python_version >= "3.12" and python_version < "3.15"
 authlib==1.6.6 ; python_version >= "3.12" and python_version < "3.15"
 certifi==2026.1.4 ; python_version >= "3.12" and python_version < "3.15"
@@ -12,14 +10,10 @@ coreason-identity==0.4.1 ; python_version >= "3.12" and python_version < "3.15"
 cryptography==46.0.4 ; python_version >= "3.12" and python_version < "3.15"
 dnspython==2.8.0 ; python_version >= "3.12" and python_version < "3.15"
 email-validator==2.3.0 ; python_version >= "3.12" and python_version < "3.15"
-fastapi==0.128.0 ; python_version >= "3.12" and python_version < "3.15"
 h11==0.16.0 ; python_version >= "3.12" and python_version < "3.15"
 httpcore==1.0.9 ; python_version >= "3.12" and python_version < "3.15"
-httpx==0.28.1 ; python_version >= "3.12" and python_version < "3.15"
 idna==3.11 ; python_version >= "3.12" and python_version < "3.15"
 importlib-metadata==8.7.1 ; python_version >= "3.12" and python_version < "3.15"
-jsonschema-specifications==2025.9.1 ; python_version >= "3.12" and python_version < "3.15"
-jsonschema==4.26.0 ; python_version >= "3.12" and python_version < "3.15"
 loguru==0.7.3 ; python_version >= "3.12" and python_version < "3.15"
 opentelemetry-api==1.39.1 ; python_version >= "3.12" and python_version < "3.15"
 pycparser==3.0 ; python_version >= "3.12" and python_version < "3.15" and platform_python_implementation != "PyPy" and implementation_name != "PyPy"
@@ -30,10 +24,7 @@ python-dotenv==1.2.1 ; python_version >= "3.12" and python_version < "3.15"
 pyyaml==6.0.3 ; python_version >= "3.12" and python_version < "3.15"
 referencing==0.37.0 ; python_version >= "3.12" and python_version < "3.15"
 rpds-py==0.30.0 ; python_version >= "3.12" and python_version < "3.15"
-starlette==0.50.0 ; python_version >= "3.12" and python_version < "3.15"
-types-aiofiles==23.2.0.20240623 ; python_version >= "3.12" and python_version < "3.15"
 typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "3.15"
 typing-inspection==0.4.2 ; python_version >= "3.12" and python_version < "3.15"
-uvicorn==0.40.0 ; python_version >= "3.12" and python_version < "3.15"
 win32-setctime==1.2.0 ; python_version >= "3.12" and python_version < "3.15" and sys_platform == "win32"
 zipp==3.23.0 ; python_version >= "3.12" and python_version < "3.15"
