+++ src/coreason_manifest/definitions/audit.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class AuditLog(BaseModel):
+    """Audit log entry."""
+
+    pass
