--- src/coreason_manifest/policies/tbom.json
@@ -1,14 +1,0 @@
-{
-  "tbom": [
-    "requests",
-    "pydantic",
-    "httpx",
-    "pandas",
-    "numpy",
-    "scipy",
-    "scikit-learn",
-    "torch",
-    "fastapi",
-    "uvicorn"
-  ]
-}
