+++ tests/test_immutability.py
@@ -4,7 +4,7 @@ import uuid
 import pytest
 from pydantic import ValidationError
 
-from coreason_manifest.models import (
+from coreason_manifest.definitions.agent import (
     AgentDefinition,
     AgentDependencies,
     ModelConfig,
