--- tests/test_integrity.py
@@ -1,172 +1,0 @@
-# Prosperity-3.0
-import os
-from pathlib import Path
-from typing import Any, Dict
-from uuid import uuid4
-
-import pytest
-
-from coreason_manifest.errors import IntegrityCompromisedError
-from coreason_manifest.integrity import IntegrityChecker
-from coreason_manifest.models import AgentDefinition
-
-
-@pytest.fixture
-def agent_def_data() -> Dict[str, Any]:
-    return {
-        "metadata": {
-            "id": str(uuid4()),
-            "version": "1.0.0",
-            "name": "Test Agent",
-            "author": "Test Author",
-            "created_at": "2023-10-26T12:00:00Z",
-        },
-        "interface": {"inputs": {}, "outputs": {}},
-        "topology": {
-            "steps": [],
-            "model_config": {"model": "gpt-4", "temperature": 0.5},
-        },
-        "dependencies": {"tools": [], "libraries": []},
-        "integrity_hash": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-    }
-
-
-def test_calculate_hash_stable(tmp_path: Path) -> None:
-    """Test that hash calculation is stable and deterministic."""
-    src = tmp_path / "src"
-    src.mkdir()
-    (src / "file1.py").write_text("print('hello')")
-    (src / "file2.py").write_text("print('world')")
-
-    hash1 = IntegrityChecker.calculate_hash(src)
-    hash2 = IntegrityChecker.calculate_hash(src)
-    assert hash1 == hash2
-
-
-def test_calculate_hash_changes(tmp_path: Path) -> None:
-    """Test that hash changes if content changes."""
-    src = tmp_path / "src"
-    src.mkdir()
-    (src / "file1.py").write_text("v1")
-
-    hash1 = IntegrityChecker.calculate_hash(src)
-
-    (src / "file1.py").write_text("v2")
-    hash2 = IntegrityChecker.calculate_hash(src)
-
-    assert hash1 != hash2
-
-
-def test_calculate_hash_structure_changes(tmp_path: Path) -> None:
-    """Test that hash changes if file name/structure changes."""
-    src = tmp_path / "src"
-    src.mkdir()
-    (src / "file1.py").write_text("v1")
-    hash1 = IntegrityChecker.calculate_hash(src)
-
-    # Rename
-    (src / "file1.py").rename(src / "file2.py")
-    hash2 = IntegrityChecker.calculate_hash(src)
-
-    assert hash1 != hash2
-
-
-def test_calculate_hash_not_found() -> None:
-    """Test Error when directory not found."""
-    with pytest.raises(FileNotFoundError):
-        IntegrityChecker.calculate_hash("non_existent")
-
-
-def test_verify_success(tmp_path: Path, agent_def_data: Dict[str, Any]) -> None:
-    """Test successful verification."""
-    src = tmp_path / "src"
-    src.mkdir()
-    (src / "main.py").write_text("pass")
-
-    expected_hash = IntegrityChecker.calculate_hash(src)
-    agent_def_data["integrity_hash"] = expected_hash
-    agent_def = AgentDefinition(**agent_def_data)
-
-    IntegrityChecker.verify(agent_def, src)
-
-
-def test_verify_mismatch(tmp_path: Path, agent_def_data: Dict[str, Any]) -> None:
-    """Test verification failure."""
-    src = tmp_path / "src"
-    src.mkdir()
-    (src / "main.py").write_text("pass")
-
-    agent_def_data["integrity_hash"] = "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
-    agent_def = AgentDefinition(**agent_def_data)
-
-    with pytest.raises(IntegrityCompromisedError) as excinfo:
-        IntegrityChecker.verify(agent_def, src)
-    assert "Integrity check failed" in str(excinfo.value)
-
-
-def test_verify_missing_hash_validation_error(tmp_path: Path, agent_def_data: Dict[str, Any]) -> None:
-    """Test verification when manifest has no hash (raises ValidationError now)."""
-    src = tmp_path / "src"
-    src.mkdir()
-    # Explicitly set to None to simulate invalid data passed to Pydantic
-    agent_def_data["integrity_hash"] = None
-    with pytest.raises(Exception) as excinfo:  # Catch Pydantic ValidationError
-        AgentDefinition(**agent_def_data)
-    assert "Input should be a valid string" in str(excinfo.value)
-
-
-def test_calculate_hash_symlink_error(tmp_path: Path) -> None:
-    """Test that symlinks raise IntegrityCompromisedError."""
-    src = tmp_path / "src"
-    src.mkdir()
-    target = src / "target.txt"
-    target.write_text("content")
-    link = src / "link.txt"
-
-    try:
-        os.symlink(target, link)
-    except OSError:
-        pytest.skip("Symlinks not supported on this platform")
-
-    with pytest.raises(IntegrityCompromisedError) as excinfo:
-        IntegrityChecker.calculate_hash(src)
-    assert "Symbolic links are forbidden" in str(excinfo.value)
-
-
-def test_calculate_hash_ignores_dirs(tmp_path: Path) -> None:
-    """Test that ignored directories are skipped."""
-    src = tmp_path / "src"
-    src.mkdir()
-    (src / "file.txt").write_text("content")
-
-    # Calculate baseline
-    hash1 = IntegrityChecker.calculate_hash(src)
-
-    # Add ignored dir
-    ignored_dir = src / ".git"
-    ignored_dir.mkdir()
-    (ignored_dir / "ignored.txt").write_text("ignored")
-
-    hash2 = IntegrityChecker.calculate_hash(src)
-
-    assert hash1 == hash2
-
-
-def test_calculate_hash_directory_symlink_error(tmp_path: Path) -> None:
-    """Test that symlinked directories raise IntegrityCompromisedError."""
-    src = tmp_path / "src"
-    src.mkdir()
-
-    # Create target outside src to ensure src only contains the link
-    target_dir = tmp_path / "target_dir"
-    target_dir.mkdir()
-
-    link_dir = src / "link_dir"
-    try:
-        os.symlink(target_dir, link_dir)
-    except OSError:
-        pytest.skip("Symlinks not supported on this platform")
-
-    with pytest.raises(IntegrityCompromisedError) as excinfo:
-        IntegrityChecker.calculate_hash(src)
-    assert "Symbolic links are forbidden" in str(excinfo.value)
