--- tests/test_main.py
@@ -1,15 +1,0 @@
-# Copyright (c) 2025 CoReason, Inc.
-#
-# This software is proprietary and dual-licensed.
-# Licensed under the Prosperity Public License 3.0 (the "License").
-# A copy of the license is available at https://prosperitylicense.com/versions/3.0.0
-# For details, see the LICENSE file.
-# Commercial use beyond a 30-day trial requires a separate license.
-#
-# Source Code: https://github.com/CoReason-AI/coreason_manifest
-
-from coreason_manifest.main import hello_world
-
-
-def test_hello_world() -> None:
-    assert hello_world() == "Hello World!"
