+++ tests/test_models_complex.py
@@ -4,7 +4,7 @@ from uuid import uuid4
 import pytest
 from pydantic import ValidationError
 
-from coreason_manifest.models import AgentMetadata
+from coreason_manifest.definitions.agent import AgentMetadata
 
 
 def test_semver_build_metadata() -> None:
