+++ tests/test_recipes_edge_cases.py
@@ -1,14 +1,9 @@
 import pytest
 from pydantic import ValidationError
 
-from coreason_manifest import (
-    AgentNode,
-    Edge,
-    GraphTopology,
-    HumanNode,
-    LogicNode,
-    RecipeManifest,
-)
+from coreason_manifest import Edge, RecipeManifest
+from coreason_manifest import Topology as GraphTopology
+from coreason_manifest.definitions.topology import AgentNode, HumanNode, LogicNode
 
 
 def test_invalid_node_discriminator() -> None:
